third_party/blink/web_tests/storage/indexeddb/createObjectStore-null-name-expected.txt

Test IndexedDB createObjectStore null handling

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "createObjectStore-null-name.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db.createObjectStore(null);
PASS objectStore.name is "null"
PASS successfullyParsed is true

TEST COMPLETE
